- remove description from profile fields - added disclaimer about DEBUG_EXTRA to the ACP (i think this is needed - some idiots might think it is wise to have this enabled on a production board. :) We *may* let it there for the Betas though, but it will be removed during the RC's) - some bugfixes
git-svn-id: file:///svn/phpbb/trunk@5973 89ea8834-ac86-4346-8a33-228a782c2dd0

+++ phpBB/includes/mcp/mcp_post.php
@@ -362,20 +362,23 @@ function change_poster(&$post_info, $userdata)
 	markread('post', $post_info['forum_id'], $post_info['topic_id'], time(), $userdata['user_id']);
 
 	// Remove the dotted topic option if the old user has no more posts within this topic
-	$sql = 'SELECT topic_id
-		FROM ' . POSTS_TABLE . '
-		WHERE topic_id = ' . $post_info['topic_id'] . '
-			AND poster_id = ' . $post_info['user_id'];
-	$result = $db->sql_query_limit($sql, 1);
-	$topic_id = (int) $db->sql_fetchfield('topic_id');
-	$db->sql_freeresult($result);
-
-	if (!$topic_id)
+	if ($config['load_db_track'] && $post_info['user_id'] != ANONYMOUS)
 	{
-		$sql = 'DELETE FROM ' . TOPICS_POSTED_TABLE . '
-			WHERE user_id = ' . $post_info['user_id'] . '
-				AND topic_id = ' . $post_info['topic_id'];
-		$db->sql_query($sql);
+		$sql = 'SELECT topic_id
+			FROM ' . POSTS_TABLE . '
+			WHERE topic_id = ' . $post_info['topic_id'] . '
+				AND poster_id = ' . $post_info['user_id'];
+		$result = $db->sql_query_limit($sql, 1);
+		$topic_id = (int) $db->sql_fetchfield('topic_id');
+		$db->sql_freeresult($result);
+
+		if (!$topic_id)
+		{
+			$sql = 'DELETE FROM ' . TOPICS_POSTED_TABLE . '
+				WHERE user_id = ' . $post_info['user_id'] . '
+					AND topic_id = ' . $post_info['topic_id'];
+			$db->sql_query($sql);
+		}
 	}
 
 	// Do not change the poster_id within the attachments table, since they were still posted by the original user
